pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "os-notif-telegram"
version = "0.1.0"
description = "Forward OS notifications to Telegram"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["notifications", "telegram", "windows", "toast", "winrt"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Topic :: Communications",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.28",
    "click>=8.0",
    "platformdirs>=3.0",
    "pystray>=0.19",
    "Pillow>=9.0",
    "winrt-runtime",
    "winrt-Windows.UI.Notifications",
    "winrt-Windows.UI.Notifications.Management",
    "winrt-Windows.Foundation",
    "winrt-Windows.Data.Xml.Dom",
]

[project.urls]
Homepage = "https://github.com/your-username/os-notif-telegram"
Issues = "https://github.com/your-username/os-notif-telegram/issues"

[project.scripts]
os-notif-telegram = "os_notif_telegram.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/os_notif_telegram"]
